modularized hooks
