+++ src/components/DialecticalWheel/hooks/index.ts
@@ -1,0 +1,8 @@
+export { useWheelSequence } from './useWheelSequence';
+export { useWheelInteraction } from './useWheelInteraction';
+export { useWheelSlices } from './useWheelSlices';
+export { useNodeConnections } from './useNodeConnections';
+
+export type { SliceConfig, SequenceWithLabels } from './useWheelSequence';
+export type { DynamicSlice, DetailedSlices, PairTexts, SliceClickData } from './useWheelSlices';
+export type { DemoConnection } from './useNodeConnections';
