+++ src/index.ts
@@ -13,13 +13,8 @@ export {
 
 // Export all utilities
 export { 
-  createSliceAtAngle,
-  generateWheelFromWisdomUnits,
-  createThesisAntithesisWheel,
-  createWisdomUnit,
-  createComponent,
+  SliceAtAngle,
   generatePairTextsFromWisdomUnits,
-  generateDetailedSlice,
   defaultPairTexts
 } from './utils';
